requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
argparse==1.2.1
gunicorn==0.16.1
pymongo==2.3
twitter-text-py==1.0.3
wsgiref==0.1.2
